tools: add nm-in-vm script
Script to download, configure and install a virtual machine to build and
test NetworkManager. This is useful because there are some things that
doesn't work properly on containers so a VM is needed to test.

It works almost the same way than nm-in-container.

Configurations specific to NetworkManager such as installing the required
packages are not implemented yet.