docs: support docblock comments in "enums-to-docbook.pl"
This script only supported parsing inline documentation comments, of the
type "@ENUM_VAL_NAME: description".

Support also docblock comments, of the type:

    typedef enum {
        /**
         * ENUM_VAL_NAME
         *
         * description
         */
         ENUM_VAL_NAME,
         ...
    }

If both styles are used for the same enum value, the docblock takes
precedence since it's probably more detailed.

Without this change, adding a docblock comment was causing an error in
this script, and then the whole build failed completely.

We need to support this because gtkdoc annotations like "(attributes
NM.internal=1)" only work in docblock comments, not in inline comments,
and we want to start using them for the "enumvals expansion" in
documentation.